Moved app.tsx to root
